[TVList]restructure TVList, add TimSort interface
