+++ server/src/main/java/org/apache/iotdb/db/utils/datastructure/DoubleTVList.java
@@ -30,22 +30,13 @@ import org.apache.iotdb.tsfile.read.common.block.TsBlockBuilder;
 import org.apache.iotdb.tsfile.utils.BitMap;
 import org.apache.iotdb.tsfile.utils.TsPrimitiveType;
 
-import java.io.DataInputStream;
-import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
 import static org.apache.iotdb.db.rescon.PrimitiveArrayManager.ARRAY_SIZE;
 
-public class DoubleTVList extends TVList {
-
-  // list of primitive array, add 1 when expanded -> double primitive array
-  // index relation: arrayIndex -> elementIndex
-  private List<double[]> values;
-
-  private double[][] sortedValues;
-
-  private double pivotValue;
+public abstract class DoubleTVList extends TVList {
+  protected List<double[]> values;
 
   DoubleTVList() {
     super();
@@ -87,40 +78,6 @@ public class DoubleTVList extends TVList {
   }
 
   @Override
-  public DoubleTVList clone() {
-    DoubleTVList cloneList = new DoubleTVList();
-    cloneAs(cloneList);
-    for (double[] valueArray : values) {
-      cloneList.values.add(cloneValue(valueArray));
-    }
-    return cloneList;
-  }
-
-  private double[] cloneValue(double[] array) {
-    double[] cloneArray = new double[array.length];
-    System.arraycopy(array, 0, cloneArray, 0, array.length);
-    return cloneArray;
-  }
-
-  @Override
-  public void sort() {
-    if (sortedTimestamps == null
-        || sortedTimestamps.length < PrimitiveArrayManager.getArrayRowCount(rowCount)) {
-      sortedTimestamps =
-          (long[][]) PrimitiveArrayManager.createDataListsByType(TSDataType.INT64, rowCount);
-    }
-    if (sortedValues == null
-        || sortedValues.length < PrimitiveArrayManager.getArrayRowCount(rowCount)) {
-      sortedValues =
-          (double[][]) PrimitiveArrayManager.createDataListsByType(TSDataType.DOUBLE, rowCount);
-    }
-    sort(0, rowCount);
-    clearSortedValue();
-    clearSortedTime();
-    sorted = true;
-  }
-
-  @Override
   void clearValue() {
     if (values != null) {
       for (double[] dataArray : values) {
@@ -131,60 +88,8 @@ public class DoubleTVList extends TVList {
   }
 
   @Override
-  void clearSortedValue() {
-    if (sortedValues != null) {
-      sortedValues = null;
-    }
-  }
-
-  @Override
-  protected void setFromSorted(int src, int dest) {
-    set(
-        dest,
-        sortedTimestamps[src / ARRAY_SIZE][src % ARRAY_SIZE],
-        sortedValues[src / ARRAY_SIZE][src % ARRAY_SIZE]);
-  }
-
-  @Override
-  protected void set(int src, int dest) {
-    long srcT = getTime(src);
-    double srcV = getDouble(src);
-    set(dest, srcT, srcV);
-  }
-
-  @Override
-  protected void setToSorted(int src, int dest) {
-    sortedTimestamps[dest / ARRAY_SIZE][dest % ARRAY_SIZE] = getTime(src);
-    sortedValues[dest / ARRAY_SIZE][dest % ARRAY_SIZE] = getDouble(src);
-  }
-
-  @Override
-  protected void reverseRange(int lo, int hi) {
-    hi--;
-    while (lo < hi) {
-      long loT = getTime(lo);
-      double loV = getDouble(lo);
-      long hiT = getTime(hi);
-      double hiV = getDouble(hi);
-      set(lo++, hiT, hiV);
-      set(hi--, loT, loV);
-    }
-  }
-
-  @Override
   protected void expandValues() {
     values.add((double[]) getPrimitiveArraysByType(TSDataType.DOUBLE));
-  }
-
-  @Override
-  protected void saveAsPivot(int pos) {
-    pivotTime = getTime(pos);
-    pivotValue = getDouble(pos);
-  }
-
-  @Override
-  protected void setPivotTo(int pos) {
-    set(pos, pivotTime, pivotValue);
   }
 
   @Override
@@ -321,18 +226,5 @@ public class DoubleTVList extends TVList {
       buffer.putLong(getTime(rowIdx));
       buffer.putDouble(getDouble(rowIdx));
     }
-  }
-
-  public static DoubleTVList deserialize(DataInputStream stream) throws IOException {
-    DoubleTVList tvList = new DoubleTVList();
-    int rowCount = stream.readInt();
-    long[] times = new long[rowCount];
-    double[] values = new double[rowCount];
-    for (int rowIdx = 0; rowIdx < rowCount; ++rowIdx) {
-      times[rowIdx] = stream.readLong();
-      values[rowIdx] = stream.readDouble();
-    }
-    tvList.putDoubles(times, values, null, 0, rowCount);
-    return tvList;
   }
 }
