+++ server/src/main/java/org/apache/iotdb/db/utils/datastructure/TVList.java
@@ -53,14 +53,11 @@ public abstract class TVList implements WALEntryValue {
   protected List<long[]> timestamps;
   protected int rowCount;
 
-  protected long[][] sortedTimestamps;
   protected boolean sorted = true;
   // record reference count of this tv list
   // currently this reference will only be increase because we can't know when to decrease it
-  protected AtomicInteger referenceCount;
-  protected long pivotTime;
   protected long minTime;
-
+  protected AtomicInteger referenceCount;
   private long version;
 
   public TVList() {
@@ -73,17 +70,17 @@ public abstract class TVList implements WALEntryValue {
   public static TVList newList(TSDataType dataType) {
     switch (dataType) {
       case TEXT:
-        return new BinaryTVList();
+        return new TimBinaryTVList();
       case FLOAT:
-        return new FloatTVList();
+        return new TimFloatTVList();
       case INT32:
-        return new IntTVList();
+        return new TimIntTVList();
       case INT64:
-        return new LongTVList();
+        return new TimLongTVList();
       case DOUBLE:
-        return new DoubleTVList();
+        return new TimDoubleTVList();
       case BOOLEAN:
-        return new BooleanTVList();
+        return new TimBooleanTVList();
       default:
         break;
     }
@@ -106,6 +103,8 @@ public abstract class TVList implements WALEntryValue {
   public boolean isSorted() {
     return sorted;
   }
+
+  public abstract void sort();
 
   public void increaseReferenceCount() {
     referenceCount.incrementAndGet();
@@ -222,8 +221,6 @@ public abstract class TVList implements WALEntryValue {
     throw new UnsupportedOperationException(ERR_DATATYPE_NOT_CONSISTENT);
   }
 
-  public abstract void sort();
-
   public long getMinTime() {
     return minTime;
   }
@@ -233,12 +230,6 @@ public abstract class TVList implements WALEntryValue {
   }
 
   protected abstract void set(int src, int dest);
-
-  protected abstract void setFromSorted(int src, int dest);
-
-  protected abstract void setToSorted(int src, int dest);
-
-  protected abstract void reverseRange(int lo, int hi);
 
   protected abstract void expandValues();
 
@@ -295,10 +286,7 @@ public abstract class TVList implements WALEntryValue {
     sorted = true;
     minTime = Long.MAX_VALUE;
     clearTime();
-    clearSortedTime();
-
     clearValue();
-    clearSortedValue();
   }
 
   protected void clearTime() {
@@ -310,19 +298,7 @@ public abstract class TVList implements WALEntryValue {
     }
   }
 
-  protected void clearSortedTime() {
-    if (sortedTimestamps != null) {
-      sortedTimestamps = null;
-    }
-  }
-
   abstract void clearValue();
-
-  /**
-   * The arrays for sorting are not including in write memory now, the memory usage is considered as
-   * temporary memory.
-   */
-  abstract void clearSortedValue();
 
   protected void checkExpansion() {
     if ((rowCount % ARRAY_SIZE) == 0) {
@@ -339,154 +315,6 @@ public abstract class TVList implements WALEntryValue {
     long[] cloneArray = new long[array.length];
     System.arraycopy(array, 0, cloneArray, 0, array.length);
     return cloneArray;
-  }
-
-  protected void sort(int lo, int hi) {
-    if (sorted) {
-      return;
-    }
-
-    if (lo == hi) {
-      return;
-    }
-    if (hi - lo <= SMALL_ARRAY_LENGTH) {
-      int initRunLen = countRunAndMakeAscending(lo, hi);
-      binarySort(lo, hi, lo + initRunLen);
-      return;
-    }
-    int mid = (lo + hi) >>> 1;
-    sort(lo, mid);
-    sort(mid, hi);
-    merge(lo, mid, hi);
-  }
-
-  protected int countRunAndMakeAscending(int lo, int hi) {
-    assert lo < hi;
-    int runHi = lo + 1;
-    if (runHi == hi) {
-      return 1;
-    }
-
-    // Find end of run, and reverse range if descending
-    if (getTime(runHi++) < getTime(lo)) { // Descending
-      while (runHi < hi && getTime(runHi) < getTime(runHi - 1)) {
-        runHi++;
-      }
-      reverseRange(lo, runHi);
-    } else { // Ascending
-      while (runHi < hi && getTime(runHi) >= getTime(runHi - 1)) {
-        runHi++;
-      }
-    }
-
-    return runHi - lo;
-  }
-
-  protected int compare(int idx1, int idx2) {
-    long t1 = getTime(idx1);
-    long t2 = getTime(idx2);
-    return Long.compare(t1, t2);
-  }
-
-  protected abstract void saveAsPivot(int pos);
-
-  protected abstract void setPivotTo(int pos);
-
-  /** From TimSort.java */
-  protected void binarySort(int lo, int hi, int start) {
-    assert lo <= start && start <= hi;
-    if (start == lo) {
-      start++;
-    }
-    for (; start < hi; start++) {
-
-      saveAsPivot(start);
-      // Set left (and right) to the index where a[start] (pivot) belongs
-      int left = lo;
-      int right = start;
-      assert left <= right;
-      /*
-       * Invariants:
-       *   pivot >= all in [lo, left).
-       *   pivot <  all in [right, start).
-       */
-      while (left < right) {
-        int mid = (left + right) >>> 1;
-        if (compare(start, mid) < 0) {
-          right = mid;
-        } else {
-          left = mid + 1;
-        }
-      }
-      assert left == right;
-
-      /*
-       * The invariants still hold: pivot >= all in [lo, left) and
-       * pivot < all in [left, start), so pivot belongs at left.  Note
-       * that if there are elements equal to pivot, left points to the
-       * first slot after them -- that's why this sort is stable.
-       * Slide elements over to make room for pivot.
-       */
-      int n = start - left; // The number of elements to move
-      for (int i = n; i >= 1; i--) {
-        set(left + i - 1, left + i);
-      }
-      setPivotTo(left);
-    }
-    for (int i = lo; i < hi; i++) {
-      setToSorted(i, i);
-    }
-  }
-
-  protected void merge(int lo, int mid, int hi) {
-    // end of sorting buffer
-    int tmpIdx = 0;
-
-    // start of unmerged parts of each sequence
-    int leftIdx = lo;
-    int rightIdx = mid;
-
-    // copy the minimum elements to sorting buffer until one sequence is exhausted
-    int endSide = 0;
-    while (endSide == 0) {
-      if (compare(leftIdx, rightIdx) <= 0) {
-        setToSorted(leftIdx, lo + tmpIdx);
-        tmpIdx++;
-        leftIdx++;
-        if (leftIdx == mid) {
-          endSide = 1;
-        }
-      } else {
-        setToSorted(rightIdx, lo + tmpIdx);
-        tmpIdx++;
-        rightIdx++;
-        if (rightIdx == hi) {
-          endSide = 2;
-        }
-      }
-    }
-
-    // copy the remaining elements of another sequence
-    int start;
-    int end;
-    if (endSide == 1) {
-      start = rightIdx;
-      end = hi;
-    } else {
-      start = leftIdx;
-      end = mid;
-    }
-    for (; start < end; start++) {
-      setToSorted(start, lo + tmpIdx);
-      tmpIdx++;
-    }
-
-    // copy from sorting buffer to the original arrays so that they can be further sorted
-    // potential speed up: change the place of sorting buffer and origin data between merge
-    // iterations
-    for (int i = lo; i < hi; i++) {
-      setFromSorted(i, i);
-    }
   }
 
   void updateMinTimeAndSorted(long[] time, int start, int end) {
@@ -563,19 +391,19 @@ public abstract class TVList implements WALEntryValue {
     TSDataType dataType = ReadWriteIOUtils.readDataType(stream);
     switch (dataType) {
       case TEXT:
-        return BinaryTVList.deserialize(stream);
+        return TimBinaryTVList.deserialize(stream);
       case FLOAT:
-        return FloatTVList.deserialize(stream);
+        return TimFloatTVList.deserialize(stream);
       case INT32:
-        return IntTVList.deserialize(stream);
+        return TimIntTVList.deserialize(stream);
       case INT64:
-        return LongTVList.deserialize(stream);
+        return TimLongTVList.deserialize(stream);
       case DOUBLE:
-        return DoubleTVList.deserialize(stream);
+        return TimDoubleTVList.deserialize(stream);
       case BOOLEAN:
-        return BooleanTVList.deserialize(stream);
+        return TimBooleanTVList.deserialize(stream);
       case VECTOR:
-        return AlignedTVList.deserialize(stream);
+        return TimAlignedTVList.deserialize(stream);
       default:
         break;
     }
